--- src/components/features/login/index.tsx
@@ -1,5 +1,0 @@
-const Login = () => {
-  return <>login</>;
-};
-
-export default Login;
