--- src/components/home/index.tsx
@@ -1,5 +1,0 @@
-const HomePage = () => {
-  return <>HomePage</>;
-};
-
-export default HomePage;
