+++ src/pages/guard/service-history/data/callback/index.ts
@@ -1,0 +1,5 @@
+export const CALL_DUMMY_DATA = Array.from({ length: 10 }, () => ({
+  date: '24.09.01',
+  name: '택시 호출하기',
+  status: '완료대기',
+}));
