+++ src/pages/guard/service-history/data/hello/index.ts
@@ -1,0 +1,5 @@
+export const HELLO_DUMMY_DATA = Array.from({ length: 10 }, () => ({
+  date: '24.08.15~24.09.01',
+  name: '김숙자',
+  status: '완료대기',
+}));
