+++ src/pages/guard/service-history/types/index.tsx
@@ -1,0 +1,6 @@
+// 데이터 타입 : 서비스 이용 내역 (보호자용 페이지 내부)
+export type HistoryItem = {
+  date: string;
+  name: string;
+  status: string;
+};
