+++ src/pages/siniddo/mypage/data/detail/index.ts
@@ -1,0 +1,6 @@
+// 10개의 더미 데이터 배열 생성
+export const DUMMY_DATA = Array.from({ length: 10 }, () => ({
+  date: '24.09.01',
+  description: '택시 호출하기',
+  points: '500p',
+}));
