--- src/routes/index.tsx
@@ -1,30 +1,0 @@
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-
-import { RouterPath } from './path';
-import MainPage from '@/pages/common/main';
-import RegisterPage from '@/pages/common/register';
-import ServiceHistoryPage from '@/pages/guard/serviceHistory';
-import MyPage from '@/pages/siniddo/mypage';
-
-const router = createBrowserRouter([
-  {
-    path: RouterPath.root,
-    element: <MainPage />,
-  },
-  {
-    path: RouterPath.register,
-    element: <RegisterPage />,
-  },
-  {
-    path: RouterPath.mypage,
-    element: <MyPage />,
-  },
-  {
-    path: RouterPath.servicehistory,
-    element: <ServiceHistoryPage />,
-  },
-]);
-
-export const Routes = () => {
-  return <RouterProvider router={router} />;
-};
