--- src/routes/path.ts
@@ -1,7 +1,0 @@
-export const RouterPath = {
-  root: '/',
-  login: '/login',
-  register: '/register',
-  mypage: `/mypage`,
-  servicehistory: `/servicehistory`,
-};
