+++ src/shared/components/features/header/index.tsx
@@ -1,0 +1,9 @@
+const Header = () => {
+  return (
+    <header>
+      <h1>Header</h1>
+    </header>
+  );
+};
+
+export default Header;
