+++ src/shared/styles/variants/index.ts
