--- src/types/index.tsx
@@ -1,6 +1,0 @@
-// 데이터 타입 : 서비스 이용 내역 (보호자용 페이지 내부)
-export interface HistoryItem {
-  date: string;
-  name: string;
-  status: string;
-}
